api/requirements.txt
flask
flask_cors
psycopg2-binary
requestsle-cloud-firestore